move models and put tests into models package
