+++ pkg/models/best_value_score_test.go
@@ -1,7 +1,7 @@
-package models_test
+package models
 
 import "testing"
 
-func Test_Shipment(t *testing.T) {
+func Test_BestValueScore(t *testing.T) {
 	t.Fatal("This test needs to be implemented!")
 }
