+++ pkg/models/traffic_distribution_list_test.go
@@ -1,4 +1,4 @@
-package models_test
+package models
 
 import "testing"
