+++ pkg/models/transportation_service_provider_test.go
@@ -1,4 +1,4 @@
-package models_test
+package models
 
 import "testing"
